chore: rename eth engine module orchestrator -> service
